Add scripts to automatically create documentation
documentation/gamsdocs.pl is perl code to automatically create
documentation from a GAMS file. An executable,
documentation/gamsdocs.exe, is created so users that do not have perl
installed can still create the documentation. The documentation in
markdown format is in documentation/documentation.md. Pandoc can be used
to convert this markdown file to html, pdf, etc. The file makedocs.bat
creates the markdown documentation and converts it to html (if pandoc is
installed).