Restructure files and directories
The structure of the directories is now different. There are two main
directories, library and project, where the name of the project
directory can be made more specific and adjusted in the configuration
file. The files within library contain the general CGE code that should
not change from project to project. When working on a project, this code
should be left as is. The code and set definitions in the project
directory can be changed. There is also an empty project_template
directory with a structure that should be used as a starting point of a
new project.