Adjust the codes for the data with non-zero taxes on export
In this new setup there are 2 types of taxes on products: (1) net taxes
paid on buying domestic and imported products, the tax revenue goes to
the government  of buyer's residence; (2) net taxes paid on export, the
tax revenue goes to the government of producer's residence. Therefore
there are two taxes paid on imported products: the one in the country of
origin, levied over the basic price in that country; and one in the
country of consumption, levied of the basic price in the country of
origin plus taxes on export.